+++ include/SigmaDS3231.hpp
@@ -1,5 +1,9 @@
 #pragma once
+#ifdef ESP8266
+#include <Esp.h>
+#else
 #include <Arduino.h>
+#endif
 #include "SigmaRTC.hpp"
 
 class SigmaDS3231 : public SigmaRTC{
