+++ platformio.ini
@@ -8,13 +8,13 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-[env:megaatmega2560]
-platform = atmelavr
-board = megaatmega2560
+[env:espressif8266]
+platform = espressif8266
+board = d1_mini
+;framework = esp8266-nonos-sdk
 framework = arduino
 monitor_speed = 115200
-upload_speed = 115200
-
+upload_speed = 921600
 lib_deps = 
 	arduino-libraries/Ethernet@^2.0.1
 	bblanchon/ArduinoJson@^6.20.1
